requirements.txt
yfinance
pandas
numpy
scikit-learn
ta
plotly
streamlit
textblob
newspaper3k
requests
beautifulsoup4
pillow
python-dotenv
python-telegram-bot --pre